+++ sources/drawing/player_mouv.c
@@ -6,31 +6,31 @@
 /*   By: mekundur <mekundur@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/26 18:55:52 by drongier          #+#    #+#             */
-/*   Updated: 2025/03/21 10:16:58 by mekundur         ###   ########.fr       */
+/*   Updated: 2025/04/26 15:46:47 by mekundur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-void	m_up(t_player *player, float cos_angle, float sin_angle, int speed)
+void	m_up(t_player *player, int cos_angle, int sin_angle, int speed)
 {
 	player->x += cos_angle * speed;
 	player->y += sin_angle * speed;
 }
 
-void	m_down(t_player *player, float cos_angle, float sin_angle, int speed)
+void	m_down(t_player *player, int cos_angle, int sin_angle, int speed)
 {
 	player->x -= cos_angle * speed;
 	player->y -= sin_angle * speed;
 }
 
-void	m_left(t_player *player, float cos_angle, float sin_angle, int speed)
+void	m_left(t_player *player, int cos_angle, int sin_angle, int speed)
 {
 	player->x += sin_angle * speed;
 	player->y -= cos_angle * speed;
 }
 
-void	m_right(t_player *player, float cos_angle, float sin_angle, int speed)
+void	m_right(t_player *player, int cos_angle, int sin_angle, int speed)
 {
 	player->x -= sin_angle * speed;
 	player->y += cos_angle * speed;
